+++ zephyr/src/net/spotflow_mqtt.c
@@ -4,8 +4,10 @@
 #include <zephyr/net/mqtt.h>
 #include <zephyr/random/random.h>
 #include <zephyr/net/socket.h>
+#include <stdbool.h>
 #include <stdint.h>
 
+#include "net/spotflow_mqtt.h"
 #include "net/spotflow_connection_helper.h"
 #include "net/spotflow_device_id.h"
 #include "net/spotflow_tls.h"
@@ -14,8 +16,13 @@
 /* should at least match MBEDTLS_SSL_MAX_CONTENT_LEN - default is 4096 */
 #define APP_MQTT_BUFFER_SIZE 4096
 
+/* Maximum size of the payload of C2D messages */
+#define C2D_PAYLOAD_BUFFER_SIZE 32
+
 #define DEFAULT_GENERAL_TIMEOUT_MSEC 500
-#define SPOTFLOW_MQTT_CBOR_TOPIC "ingest-cbor"
+#define SPOTFLOW_MQTT_INGEST_CBOR_TOPIC "ingest-cbor"
+#define SPOTFLOW_MQTT_CONFIG_CBOR_D2C_TOPIC "config-cbor-d2c"
+#define SPOTFLOW_MQTT_CONFIG_CBOR_C2D_TOPIC "config-cbor-c2d"
 
 #define RC_STR(rc) ((rc) == 0 ? "OK" : "ERROR")
 
@@ -29,7 +36,9 @@ struct mqtt_config {
 	struct zsock_addrinfo* server_addr;
 	struct mqtt_utf8 username;
 	struct mqtt_utf8 password;
-	struct mqtt_utf8 topic;
+	struct mqtt_utf8 ingest_topic;
+	struct mqtt_utf8 config_d2c_topic;
+	struct mqtt_utf8 config_c2d_topic;
 };
 
 struct mqtt_client_toolset {
@@ -38,12 +47,16 @@ struct mqtt_client_toolset {
 	struct zsock_pollfd fds[1];
 	int nfds;
 	bool mqtt_connected;
+	spotflow_mqtt_message_cb c2d_message_callback;
+	uint16_t c2d_sub_message_id;
 };
 
 static int client_init(struct mqtt_client* client);
 static int poll_with_timeout(int timeout);
 static int prepare_fds();
+static int spotflow_mqtt_publish_cbor_msg(uint8_t* payload, size_t len, struct mqtt_utf8 topic);
 static void mqtt_evt_handler(struct mqtt_client* client, const struct mqtt_evt* evt);
+static bool utf8_starts_with(const struct mqtt_utf8* str, const struct mqtt_utf8* prefix);
 static void clear_fds(void);
 
 static struct mqtt_config spotflow_mqtt_config = {
@@ -52,7 +65,9 @@ static struct mqtt_config spotflow_mqtt_config = {
 	.server_addr = NULL,
 	.username = { 0 },
 	.password = MQTT_UTF8_LITERAL(CONFIG_SPOTFLOW_INGEST_KEY),
-	.topic = MQTT_UTF8_LITERAL(SPOTFLOW_MQTT_CBOR_TOPIC),
+	.ingest_topic = MQTT_UTF8_LITERAL(SPOTFLOW_MQTT_INGEST_CBOR_TOPIC),
+	.config_d2c_topic = MQTT_UTF8_LITERAL(SPOTFLOW_MQTT_CONFIG_CBOR_D2C_TOPIC),
+	.config_c2d_topic = MQTT_UTF8_LITERAL(SPOTFLOW_MQTT_CONFIG_CBOR_C2D_TOPIC),
 };
 
 static struct mqtt_client_toolset mqtt_client_toolset = { .mqtt_connected = false };
@@ -60,6 +75,9 @@ static struct mqtt_client_toolset mqtt_client_toolset = { .mqtt_connected = fals
 /* Buffers for MQTT client. */
 static uint8_t rx_buffer[APP_MQTT_BUFFER_SIZE];
 static uint8_t tx_buffer[APP_MQTT_BUFFER_SIZE];
+
+/* Buffer for C2D messages */
+static uint8_t c2d_payload_buffer[C2D_PAYLOAD_BUFFER_SIZE];
 
 int spotflow_mqtt_poll()
 {
@@ -105,6 +123,9 @@ int spotflow_mqtt_send_live()
 
 void spotflow_mqtt_establish_mqtt()
 {
+	mqtt_client_toolset.c2d_message_callback = NULL;
+	mqtt_client_toolset.c2d_sub_message_id = 0;
+
 	/* infinitely try to connect to mqtt broker */
 	while (!mqtt_client_toolset.mqtt_connected) {
 		int rc;
@@ -230,13 +251,45 @@ static int poll_with_timeout(int timeout)
 	return ret;
 }
 
-int spotflow_mqtt_publish_cbor_msg(uint8_t* payload, size_t len)
+int spotflow_mqtt_request_config_subscription(spotflow_mqtt_message_cb callback)
+{
+	mqtt_client_toolset.c2d_message_callback = callback;
+
+	struct mqtt_topic topics[] = {
+		{
+		    .topic = spotflow_mqtt_config.config_c2d_topic,
+		    .qos = MQTT_QOS_0_AT_MOST_ONCE,
+		},
+	};
+
+	struct mqtt_subscription_list param = {
+		.list = topics,
+		.list_count = ARRAY_SIZE(topics),
+		.message_id = sys_rand16_get(),
+	};
+
+	mqtt_client_toolset.c2d_sub_message_id = param.message_id;
+
+	return mqtt_subscribe(&mqtt_client_toolset.mqtt_client, &param);
+}
+
+int spotflow_mqtt_publish_ingest_cbor_msg(uint8_t* payload, size_t len)
+{
+	return spotflow_mqtt_publish_cbor_msg(payload, len, spotflow_mqtt_config.ingest_topic);
+}
+
+int spotflow_mqtt_publish_config_cbor_msg(uint8_t* payload, size_t len)
+{
+	return spotflow_mqtt_publish_cbor_msg(payload, len, spotflow_mqtt_config.config_d2c_topic);
+}
+
+static int spotflow_mqtt_publish_cbor_msg(uint8_t* payload, size_t len, struct mqtt_utf8 topic)
 {
 	struct mqtt_publish_param param;
 	/* using lowest guarantee because handling puback (for better guarantees)
 	 * is not implemented now */
 	param.message.topic.qos = MQTT_QOS_0_AT_MOST_ONCE;
-	param.message.topic.topic = spotflow_mqtt_config.topic;
+	param.message.topic.topic = topic;
 	param.message.payload.data = payload;
 	param.message.payload.len = len;
 	param.message_id = sys_rand16_get();
@@ -247,10 +300,13 @@ int spotflow_mqtt_publish_cbor_msg(uint8_t* payload, size_t len)
 
 static void mqtt_evt_handler(struct mqtt_client* client, const struct mqtt_evt* evt)
 {
-	int err;
+	int ret;
 	switch (evt->type) {
 	case MQTT_EVT_SUBACK:
 		LOG_DBG("SUBACK packet id: %u", evt->param.suback.message_id);
+		if (evt->param.suback.message_id == mqtt_client_toolset.c2d_sub_message_id) {
+			LOG_DBG("Subscription to desired configuration topic acknowledged");
+		}
 		break;
 	case MQTT_EVT_UNSUBACK:
 		LOG_DBG("UNSUBACK packet id: %u", evt->param.suback.message_id);
@@ -282,9 +338,9 @@ static void mqtt_evt_handler(struct mqtt_client* client, const struct mqtt_evt* 
 		LOG_DBG("PUBREC packet id: %u", evt->param.pubrec.message_id);
 		const struct mqtt_pubrel_param rel_param = { .message_id =
 								 evt->param.pubrec.message_id };
-		err = mqtt_publish_qos2_release(client, &rel_param);
-		if (err != 0) {
-			LOG_ERR("Failed to send MQTT PUBREL: %d", err);
+		ret = mqtt_publish_qos2_release(client, &rel_param);
+		if (ret < 0) {
+			LOG_ERR("Failed to send MQTT PUBREL: %d", ret);
 		}
 		break;
 	case MQTT_EVT_PUBCOMP:
@@ -294,9 +350,35 @@ static void mqtt_evt_handler(struct mqtt_client* client, const struct mqtt_evt* 
 		}
 		LOG_DBG("PUBCOMP packet id: %u", evt->param.pubcomp.message_id);
 		break;
+	case MQTT_EVT_PUBLISH:
+		if (evt->result != 0) {
+			LOG_ERR("MQTT PUBLISH decode error: %d", evt->result);
+			break;
+		}
+		LOG_DBG("PUBLISH packet id: %u", evt->param.publish.message_id);
+
+		/* The actual topic name is longer to distinguish between different devices */
+		if (mqtt_client_toolset.c2d_message_callback &&
+		    utf8_starts_with(&evt->param.publish.message.topic.topic,
+				     &spotflow_mqtt_config.config_c2d_topic)) {
+			ret = mqtt_read_publish_payload(client, c2d_payload_buffer,
+							sizeof(c2d_payload_buffer));
+			if (ret < 0) {
+				LOG_ERR("Failed to read PUBLISH payload: %d", ret);
+				break;
+			}
+
+			mqtt_client_toolset.c2d_message_callback(c2d_payload_buffer, ret);
+		}
+		break;
 	default:
 		break;
 	}
+}
+
+static bool utf8_starts_with(const struct mqtt_utf8* str, const struct mqtt_utf8* prefix)
+{
+	return str->size >= prefix->size && memcmp(str->utf8, prefix->utf8, prefix->size) == 0;
 }
 
 static void clear_fds(void)